Changes to `Home`
