+++ src/app/components/Item/ItemsContainer.jsx
@@ -275,5 +275,3 @@ export default {
   ItemsContainer: connect(mapStateToProps)(ItemsContainer),
   unwrappedItemsContainer: ItemsContainer,
 };
-
-// export default connect(mapStateToProps)(ItemsContainer);
