+++ src/client/styles/components/Home.scss
@@ -4,6 +4,7 @@
   }
 
   .nypl-lead {
-    font-size: 20px;
+    font-size: 16px;
+    font-weight: 300;
   }
 }
