+++ test/unit/Home.test.js
@@ -4,7 +4,7 @@ import React from 'react';
 import { expect } from 'chai';
 import { shallow } from 'enzyme';
 
-import Home from '../../src/app/components/Home/Home';
+import Home from '../../src/app/pages/Home';
 
 describe('Home', () => {
   let component;
@@ -15,26 +15,14 @@ describe('Home', () => {
 
   it('should be wrapped in a .home class', () => {
     expect(component.find('.home').length).to.equal(1);
-    expect(component.find('div').first().hasClass('home')).to.equal(true);
   });
 
-  it('should render a hero-type banner page title', () => {
-    const pageHeader = component.find('.nypl-homepage-hero');
-
-    expect(pageHeader).to.have.length(1);
-    expect(pageHeader.find('h1')).to.have.length(1);
-    expect(pageHeader.find('h1').text()).to.equal('Shared Collection Catalog');
-    expect(pageHeader.contains(<h1>Shared Collection Catalog</h1>)).to.equal(true);
-  });
-
-  xit('should contains a Search component in the banner', () => {
-    const pageHeader = component.find('.nypl-homepage-hero');
-    expect(pageHeader.find('Search')).to.have.length(1);
+  it('should contain a Search component in the banner', () => {
+    expect(component.find('Search')).to.have.length(1);
   });
 
   it('should contain an h2', () => {
-    const h2 = component.find('h2');
-    expect(h2.length).to.equal(1);
+    const h2 = component.find('Heading').at(0).dive();
     expect(h2.text()).to.equal('Research at NYPL');
   });
 
@@ -47,7 +35,8 @@ describe('Home', () => {
 
   it('should have five h3s in the image blocks', () => {
     const imageBlocks = component.find('div.nypl-quarter-image');
-
-    expect(imageBlocks.find('h3').length).to.equal(5);
+    imageBlocks.forEach((imageBlock) => {
+      expect(imageBlock.render().find('h4').length).to.equal(1);
+    });
   });
 });
